feat : 구장매니저 일정 가져오기

+++ FootballMatchingPlatform/src/main/java/com/kosta/project/controller/FieldApplyRestController.java
@@ -17,7 +17,7 @@ public class FieldApplyRestController {
 	private final SystemManagerFieldService smfs;
 	
 	@PostMapping("/api/update/approve/date")
-	public Map<String, String> updateApprove(@PathVariable Long fieldSeq){
+	public Map<String, String> updateApprove(@PathVariable int fieldSeq){
 		
 		smfs.updateFieldStatus2(fieldSeq);
